requirements.txt
kafka-python
numpy
tensorflow
Pillow